Move PythonContainerTasks to tasks package
